moving conf interfaces closer of where they are used
