+++ src/event-store/config/event-store-module-async-config.interface.ts
