+++ src/event-store/config/index.ts
@@ -1,0 +1,3 @@
+export * from './event-store-module-async-config.interface';
+export * from './event-store-config.interface';
+export * from './event-store-service-config.interface';
