+++ src/event-store/index.ts
@@ -1,4 +1,4 @@
-export * from './event-store.publisher';
+export * from './publisher/event-store.publisher';
 export * from './event-store.service';
 export * from './connector/implementations/tcp-http/tcp-http-event-store';
 export * from './event-store-aggregate-root';
