+++ src/event-store/publisher/event-store.publisher.spec.ts
@@ -1,12 +1,12 @@
 import {EventStorePublisher} from './event-store.publisher';
-import {IWriteEvent, IWriteEventBusConfig} from '../interfaces';
-import {TcpHttpEventStore} from './connector/implementations/tcp-http/tcp-http-event-store';
+import {IWriteEvent, IWriteEventBusConfig} from '../../interfaces';
+import {TcpHttpEventStore} from '../connector/implementations/tcp-http/tcp-http-event-store';
 import {of} from 'rxjs';
-import {ExpectedVersion} from '../enum';
+import {ExpectedVersion} from '../../enum';
 import spyOn = jest.spyOn;
-import {EventStoreService} from './event-store.service';
+import {EventStoreService} from '../event-store.service';
 
-jest.mock('./connector/implementations/tcp-http/tcp-http-event-store');
+jest.mock('../connector/implementations/tcp-http/tcp-http-event-store');
 
 describe('EventStorePublisher', () => {
     let publisher: EventStorePublisher;
