+++ src/interfaces/config/index.ts
@@ -1,7 +1,4 @@
 export * from './event-bus-config.type';
-export * from './event-store-module-async-config.interface';
-export * from './event-store-config.interface';
-export * from './event-store-service-config.interface';
 export * from './event-bus-prepublish-config.interface';
 export * from './read-event-bus-config.type';
 export * from './write-event-bus-config.interface';
